+++ PlatypusTools.UI/ViewModels/SystemRestoreViewModel.cs
@@ -30,7 +30,7 @@ namespace PlatypusTools.UI.ViewModels
 
         public SystemRestoreViewModel()
         {
-            _systemRestoreService = new SystemRestoreService();
+            _systemRestoreService = Services.ServiceLocator.SystemRestore;
 
             RefreshCommand = new RelayCommand(async _ => await RefreshAsync(), _ => !IsLoading);
             CreatePointCommand = new RelayCommand(async _ => await CreatePointAsync(), _ => !IsLoading);